refactor: split schema types into separate files
This consolidates all related methods into the same file per type,
making changes quite a bit simpler.
