+++ backend/schema/schema.go
@@ -1,259 +1,45 @@
 package schema
 
 import (
+	"crypto/sha256"
 	"fmt"
-	"io"
-	"reflect"
-	"sort"
 	"strings"
 
-	"github.com/alecthomas/participle/v2"
-	"github.com/alecthomas/participle/v2/lexer"
 	"github.com/alecthomas/types"
-	"golang.org/x/exp/maps"
 	"google.golang.org/protobuf/proto"
+
+	schemapb "github.com/TBD54566975/ftl/protos/xyz/block/ftl/v1/schema"
 )
 
-var (
-	declUnion            = []Decl{&Data{}, &Verb{}}
-	nonOptionalTypeUnion = []Type{&Int{}, &Float{}, &String{}, &Bytes{}, &Bool{}, &Time{}, &Array{}, &Map{} /*&VerbRef{},*/, &DataRef{}}
-	typeUnion            = append(nonOptionalTypeUnion, &Optional{})
-	metadataUnion        = []Metadata{&MetadataCalls{}, &MetadataIngress{}}
-	ingressUnion         = []IngressPathComponent{&IngressPathLiteral{}, &IngressPathParameter{}}
-
-	// Used by protobuf generation.
-	unions = map[reflect.Type][]reflect.Type{
-		reflect.TypeOf((*Type)(nil)).Elem():                 reflectUnion(typeUnion...),
-		reflect.TypeOf((*Metadata)(nil)).Elem():             reflectUnion(metadataUnion...),
-		reflect.TypeOf((*IngressPathComponent)(nil)).Elem(): reflectUnion(ingressUnion...),
-		reflect.TypeOf((*Decl)(nil)).Elem():                 reflectUnion(declUnion...),
-	}
-)
-
-type Position struct {
-	Filename string `protobuf:"1"`
-	Offset   int    `parser:"" protobuf:"-"`
-	Line     int    `protobuf:"2"`
-	Column   int    `protobuf:"3"`
-}
-
-func (p Position) String() string {
-	if p.Filename == "" {
-		return fmt.Sprintf("%d:%d", p.Line, p.Column)
-	}
-	return fmt.Sprintf("%s:%d:%d", p.Filename, p.Line, p.Column)
-}
-
-// A Node in the schema grammar.
-//
-//sumtype:decl
-type Node interface {
-	String() string
-	ToProto() proto.Message
-	// schemaChildren returns the children of this node.
-	schemaChildren() []Node
-}
-
-// Type represents a Type Node in the schema grammar.
-//
-//sumtype:decl
-type Type interface {
-	Node
-	// schemaType is a marker to ensure that all sqltypes implement the Type interface.
-	schemaType()
-}
-
-// Optional represents a Type whose value may be optional.
-type Optional struct {
+type Schema struct {
 	Pos Position `parser:"" protobuf:"1,optional"`
 
-	Type Type `parser:"@@" protobuf:"2,optional"`
+	Modules []*Module `parser:"@@*" protobuf:"2"`
 }
 
-type Int struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
+var _ Node = (*Schema)(nil)
 
-	Int bool `parser:"@'Int'" protobuf:"-"`
-}
-
-type Float struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Float bool `parser:"@'Float'" protobuf:"-"`
-}
-
-type String struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Str bool `parser:"@'String'" protobuf:"-"`
-}
-
-type Bytes struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Bytes bool `parser:"@'Bytes'" protobuf:"-"`
-}
-
-type Bool struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Bool bool `parser:"@'Bool'" protobuf:"-"`
-}
-
-type Time struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Time bool `parser:"@'Time'" protobuf:"-"`
-}
-
-type Array struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Element Type `parser:"'[' @@ ']'" protobuf:"2"`
-}
-
-type Map struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Key   Type `parser:"'{' @@" protobuf:"2"`
-	Value Type `parser:"':' @@ '}'" protobuf:"3"`
-}
-
-type Field struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Comments []string `parser:"@Comment*" protobuf:"3"`
-	Name     string   `parser:"@Ident" protobuf:"2"`
-	Type     Type     `parser:"@@" protobuf:"4"`
-}
-
-// Ref is a reference to another symbol.
-type Ref struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Module string `parser:"(@Ident '.')?" protobuf:"3"`
-	Name   string `parser:"@Ident" protobuf:"2"`
-}
-
-func (r *Ref) String() string {
-	return makeRef(r.Module, r.Name)
-}
-
-// DataRef is a reference to a data structure.
-type DataRef Ref
-
-// A Data structure.
-type Data struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Comments []string   `parser:"@Comment*" protobuf:"5"`
-	Name     string     `parser:"'data' @Ident '{'" protobuf:"2"`
-	Fields   []*Field   `parser:"@@* '}'" protobuf:"3"`
-	Metadata []Metadata `parser:"@@*" protobuf:"4"`
-}
-
-// VerbRef is a reference to a Verb.
-type VerbRef Ref
-
-type Verb struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Comments []string   `parser:"@Comment*" protobuf:"3"`
-	Name     string     `parser:"'verb' @Ident" protobuf:"2"`
-	Request  *DataRef   `parser:"'(' @@ ')'" protobuf:"4"`
-	Response *DataRef   `parser:"@@" protobuf:"5"`
-	Metadata []Metadata `parser:"@@*" protobuf:"6"`
-}
-
-// AddCall adds a call reference to the Verb.
-func (v *Verb) AddCall(verb *VerbRef) {
-	for _, c := range v.Metadata {
-		if c, ok := c.(*MetadataCalls); ok {
-			c.Calls = append(c.Calls, verb)
-			return
+func (s *Schema) String() string {
+	out := &strings.Builder{}
+	for i, m := range s.Modules {
+		if i != 0 {
+			fmt.Fprintln(out)
 		}
+		fmt.Fprint(out, m)
 	}
-	v.Metadata = append(v.Metadata, &MetadataCalls{Calls: []*VerbRef{verb}})
+	return out.String()
 }
 
-type Metadata interface {
-	Node
-	schemaMetadata()
-}
-
-type MetadataCalls struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Calls []*VerbRef `parser:"'calls' @@ (',' @@)*" protobuf:"2"`
-}
-
-type MetadataIngress struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Method string                 `parser:"'ingress' @('GET' | 'POST')" protobuf:"2"`
-	Path   []IngressPathComponent `parser:"('/' @@)+" protobuf:"3"`
-}
-
-type IngressPathComponent interface {
-	Node
-	schemaIngressPathComponent()
-}
-
-type IngressPathLiteral struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Text string `parser:"@Ident" protobuf:"2"`
-}
-
-type IngressPathParameter struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Name string `parser:"'{' @Ident '}'" protobuf:"2"`
-}
-
-type Module struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Comments []string `parser:"@Comment*" protobuf:"2"`
-	Builtin  bool     `parser:"@'builtin'?" protobuf:"3"`
-	Name     string   `parser:"'module' @Ident '{'" protobuf:"4"`
-	Decls    []Decl   `parser:"@@* '}'" protobuf:"5"`
-}
-
-type Decl interface {
-	Node
-	schemaDecl()
-}
-
-// AddData and return its index.
-func (m *Module) AddData(data *Data) int {
-	for i, d := range m.Decls {
-		if d, ok := d.(*Data); ok && d.Name == data.Name {
-			return i
-		}
+func (s *Schema) schemaChildren() []Node {
+	out := make([]Node, len(s.Modules))
+	for i, m := range s.Modules {
+		out[i] = m
 	}
-	m.Decls = append(m.Decls, data)
-	return len(m.Decls) - 1
+	return out
 }
 
-func (m *Module) Verbs() []*Verb {
-	var verbs []*Verb
-	for _, d := range m.Decls {
-		if v, ok := d.(*Verb); ok {
-			verbs = append(verbs, v)
-		}
-	}
-	return verbs
-}
-
-func (m *Module) Data() []*Data {
-	var data []*Data
-	for _, d := range m.Decls {
-		if v, ok := d.(*Data); ok {
-			data = append(data, v)
-		}
-	}
-	return data
+func (s *Schema) Hash() [sha256.Size]byte {
+	return sha256.Sum256([]byte(s.String()))
 }
 
 func (s *Schema) ResolveDataRef(ref *DataRef) *Data {
@@ -280,37 +66,6 @@ func (s *Schema) ResolveVerbRef(ref *VerbRef) *Verb {
 		}
 	}
 	return nil
-}
-
-// Imports returns the modules imported by this module.
-func (m *Module) Imports() []string {
-	imports := map[string]bool{}
-	_ = Visit(m, func(n Node, next func() error) error {
-		switch n := n.(type) {
-		case *DataRef:
-			if n.Module != "" && n.Module != m.Name {
-				imports[n.Module] = true
-			}
-
-		case *VerbRef:
-			if n.Module != "" && n.Module != m.Name {
-				imports[n.Module] = true
-			}
-
-		default:
-		}
-		return next()
-	})
-
-	importStrs := maps.Keys(imports)
-	sort.Strings(importStrs)
-	return importStrs
-}
-
-type Schema struct {
-	Pos Position `parser:"" protobuf:"1,optional"`
-
-	Modules []*Module `parser:"@@*" protobuf:"2"`
 }
 
 // Module returns the named module if it exists.
@@ -346,89 +101,21 @@ func (s *Schema) Upsert(module *Module) {
 	s.Modules = append(s.Modules, module)
 }
 
-var (
-	lex = lexer.MustSimple([]lexer.SimpleRule{
-		{Name: "Whitespace", Pattern: `\s+`},
-		{Name: "Ident", Pattern: `\b[a-zA-Z_][a-zA-Z0-9_]*\b`},
-		{Name: "Comment", Pattern: `//.*`},
-		{Name: "String", Pattern: `"(?:\\.|[^"])*"`},
-		{Name: "Number", Pattern: `[0-9]+(?:\.[0-9]+)?`},
-		{Name: "Punct", Pattern: `[/-:[\]{}<>()*+?.,\\^$|#]`},
-	})
-
-	commonParserOptions = []participle.Option{
-		participle.Lexer(lex),
-		participle.Elide("Whitespace"),
-		participle.Unquote(),
-		participle.Map(func(token lexer.Token) (lexer.Token, error) {
-			token.Value = strings.TrimSpace(strings.TrimPrefix(token.Value, "//"))
-			return token, nil
-		}, "Comment"),
-		participle.Union(metadataUnion...),
-		participle.Union(ingressUnion...),
-		participle.Union(declUnion...),
+func (s *Schema) ToProto() proto.Message {
+	return &schemapb.Schema{
+		Pos:     posToProto(s.Pos),
+		Modules: nodeListToProto[*schemapb.Module](s.Modules),
 	}
-
-	// Parser options for every parser _except_ the type parser.
-	parserOptions = append(commonParserOptions, participle.ParseTypeWith(parseType))
-
-	parser       = participle.MustBuild[Schema](parserOptions...)
-	moduleParser = participle.MustBuild[Module](parserOptions...)
-	refParser    = participle.MustBuild[Ref](parserOptions...)
-	typeParser   = participle.MustBuild[typeParserGrammar](append(commonParserOptions, participle.Union(nonOptionalTypeUnion...))...)
-)
-
-// We have a separate parser for types because Participle doesn't support left
-// recursion and "Type = Type ? | Int | String ..." is left recursive.
-type typeParserGrammar struct {
-	Type     Type `parser:"@@"`
-	Optional bool `parser:"@'?'?"`
 }
 
-func parseType(pl *lexer.PeekingLexer) (Type, error) {
-	typ, err := typeParser.ParseFromLexer(pl, participle.AllowTrailing(true))
+// FromProto converts a protobuf Schema to a Schema and validates it.
+func FromProto(s *schemapb.Schema) (*Schema, error) {
+	modules, err := moduleListToSchema(s.Modules)
 	if err != nil {
 		return nil, err
 	}
-	if typ.Optional {
-		return &Optional{Type: typ.Type}, nil
+	schema := &Schema{
+		Modules: modules,
 	}
-	return typ.Type, nil
-}
-
-func ParseString(filename, input string) (*Schema, error) {
-	mod, err := parser.ParseString(filename, input)
-	if err != nil {
-		return nil, err
-	}
-	return Validate(mod)
-}
-
-func ParseModuleString(filename, input string) (*Module, error) {
-	mod, err := moduleParser.ParseString(filename, input)
-	if err != nil {
-		return nil, err
-	}
-	return mod, ValidateModule(mod)
-}
-
-func ParseRef(ref string) (*Ref, error) {
-	r, err := refParser.ParseString("", ref)
-	return r, err
-}
-
-func Parse(filename string, r io.Reader) (*Schema, error) {
-	mod, err := parser.Parse(filename, r)
-	if err != nil {
-		return nil, err
-	}
-	return Validate(mod)
-}
-
-func ParseModule(filename string, r io.Reader) (*Module, error) {
-	mod, err := moduleParser.Parse(filename, r)
-	if err != nil {
-		return nil, err
-	}
-	return mod, ValidateModule(mod)
+	return Validate(schema)
 }
